Add Handlebars processing Add KV configuration Update documentation Split rules into their own files Update test cases
