+++ package.json
@@ -55,6 +55,7 @@
     "@oclif/command": "^1.8.0",
     "chalk": "^4.1.0",
     "cli-ux": "^5.5.1",
+    "handlebars": "^4.7.7",
     "jmespath": "^0.15.0"
   },
   "oclif": {
